+++ runner/meta/src/mill/runner/meta/CliImports.scala
@@ -1,5 +1,4 @@
-package mill.runner
-
+package mill.runner.meta
 import scala.util.DynamicVariable
 
 /**
